# docs1/swagger.yaml
basePath: /
definitions:
...
  handler.createListV2Request:
    properties:
      description:
        type: string
      idempotency_key:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.getAllListsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/todo.TodoList'
        type: array
    type: object
  handler.getAllListsV2Response:
    properties:
      data:
        items:
          $ref: '#/definitions/todo.TodoList'
        type: array
      meta:
        $ref: '#/definitions/handler.paginationMeta'
    type: object
...
  handler.idempotentResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
  handler.paginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
...
  handler.todoListV2Response:
    properties:
      archived:
        description: Новое поле для v2
        type: boolean
      color:
        description: Новое поле в v2
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      item_count:
        type: integer
      priority:
        description: Новое поле в v2
        type: integer
      title:
        type: string
      updated_at:
        type: string
    required:
    - title
    type: object
  handler.updateListV2Request:
    properties:
      archived:
        type: boolean
      description:
        type: string
      title:
        type: string
    type: object
...
  todo.TodoList:
    properties:
      archived:
        description: Новое поле для v2
        type: boolean
      color:
        description: Новое поле в v2
        type: string
      created_at:
        description: Новое поле для v2
        type: string
      description:
        type: string
      id:
        type: integer
      priority:
        description: Новое поле в v2
        type: integer
      title:
        type: string
      updated_at:
        description: Новое поле для v2
        type: string
    required:
    - title
    type: object
...
  todo.UpdateListInput:
    properties:
      archived:
        description: Добавлено для v2
        type: boolean
      color:
        description: Добавлено для v2
        type: string
      description:
        type: string
      priority:
        description: Добавлено для v2
        type: integer
      title:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for TodoList Application with multiple versions, rate limiting
    and idempotency support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Todo App API
  version: "2.0"
paths:
...
  /api/v1/lists:
    get:
      consumes:
      - application/json
      description: get all lists
      operationId: get-all-lists
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllListsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all lists
      tags:
      - lists
    post:
      consumes:
      - application/json
      description: create todo list
      operationId: create-list
      parameters:
      - description: list info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo.TodoList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create todo list
      tags:
      - lists
  /api/v1/lists/{id}:
    delete:
      consumes:
      - application/json
      description: delete list
      operationId: delete-list
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete list
      tags:
      - lists
    get:
      consumes:
      - application/json
      description: get list by ID
      operationId: get-list-by-id
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todo.TodoList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get list by ID
      tags:
      - lists
    put:
      consumes:
      - application/json
      description: update list
      operationId: update-list
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        type: integer
      - description: List update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/todo.UpdateListInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update list
      tags:
      - lists
...
  /api/v2/lists:
    get:
      consumes:
      - application/json
      description: Get all todo lists with pagination and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by archived status
        in: query
        name: archived
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllListsV2Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all lists (v2)
      tags:
      - lists-v2
    post:
      consumes:
      - application/json
      description: Create todo list with idempotency support
      parameters:
      - description: List info with idempotency key
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.createListV2Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.idempotentResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/todo.TodoList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create todo list (v2)
      tags:
      - lists-v2
  /api/v2/lists/{id}:
    delete:
      consumes:
      - application/json
      description: Delete todo list with soft delete
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete list (v2)
      tags:
      - lists-v2
    get:
      consumes:
      - application/json
      description: Get todo list by ID with extended information including item count
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.todoListV2Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get list by ID (v2)
      tags:
      - lists-v2
    put:
      consumes:
      - application/json
      description: Update todo list with partial update support
      parameters:
      - description: List ID
        in: path
        name: id
        required: true
        type: integer
      - description: List update data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.updateListV2Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todo.TodoList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update list (v2)
      tags:
      - lists-v2
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...